flask/requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
python-dotenv
pytz
pandas 
geopandas 
folium 
scikit-learn
joblib
Flask-Cors
requests
pycuda
numpy